Move directory path implementations into DirectoryPath hierarchy to make directory semantics explicit
