+++ src/Path/Directory/NormalizedDirectoryPath.php
@@ -2,14 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Path;
+namespace Haspadar\Piqule\Path\Directory;
 
+use Haspadar\Piqule\Path\Path;
 use Override;
 
 final readonly class NormalizedDirectoryPath implements Path
 {
     public function __construct(
-        private DirectoryPath $origin,
+        private AbsoluteDirectoryPath $origin,
     ) {}
 
     #[Override]
