+++ src/Path/Directory/ValidatedDirectoryPath.php
@@ -2,15 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Path;
+namespace Haspadar\Piqule\Path\Directory;
 
+use Haspadar\Piqule\Path\Path;
 use Haspadar\Piqule\PiquleException;
 use Override;
 
 final readonly class ValidatedDirectoryPath implements Path
 {
     public function __construct(
-        private DirectoryPath $origin,
+        private AbsoluteDirectoryPath $origin,
     ) {}
 
     #[Override]
