+++ tests/Unit/Path/Directory/AbsoluteDirectoryPathTest.php
@@ -2,20 +2,20 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\Tests\Unit\Path;
+namespace Haspadar\Piqule\Tests\Unit\Path\Directory;
 
-use Haspadar\Piqule\Path\DirectoryPath;
+use Haspadar\Piqule\Path\Directory\AbsoluteDirectoryPath;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\TestCase;
 
-final class DirectoryPathTest extends TestCase
+final class AbsoluteDirectoryPathTest extends TestCase
 {
     #[Test]
     public function returnsRawValue(): void
     {
         self::assertSame(
             'any/path/value',
-            (new DirectoryPath('any/path/value'))->value(),
+            (new AbsoluteDirectoryPath('any/path/value'))->value(),
             'Returns raw directory path value as provided',
         );
     }
@@ -23,12 +23,12 @@ final class DirectoryPathTest extends TestCase
     #[Test]
     public function returnsPosixRoot(): void
     {
-        self::assertSame('/', (new DirectoryPath('/'))->value());
+        self::assertSame('/', (new AbsoluteDirectoryPath('/'))->value());
     }
 
     #[Test]
     public function returnsWindowsDriveRoot(): void
     {
-        self::assertSame('C:\\', (new DirectoryPath('C:\\'))->value());
+        self::assertSame('C:\\', (new AbsoluteDirectoryPath('C:\\'))->value());
     }
 }
